Replace NodeOperation with NodeTypeOperation
NodeTypeOperation will refer to objects which act on groups of nodes, i.e. all nodes of a certain type

NodeOperation will refer to objects which act only on some subset of the nodes of a given type, e.g. a pulling force will only pull on the top 10% of membrane nodes.